Fallback example fix for Internet Explorer
The intention of this example seems to be to illustrate how the fallback is enabled in browsers that don't support input type="time". The statement test.type = 'time'; fails in Internet Explorer 11, the obvious example of a browser that doesn't support type="time". I've changed the example so that it works in the Internet Explorer.