<ul> was/is forgotten
In the example given https://developer.mozilla.org/en-US/docs/Learn/HTML/Introduction_to_HTML/The_head_metadata_in_HTML the image shows there is a list. With the given code to use for example, it does not include the unordered list and its content. I am not sure how to properly submit a request as I am completely new to this, but I hope that it reaches you well. 
Thank you!