Update contact-search.html
if the 'contact not found' code is inside the for loop, for each contact not the one we want, the text 'Contact not found.' would be shown, even if the contact we want is found latterly. 
Though the text 'Contact not found.' is shown for a very short time as long as finally we found the contact we want. But the logic could be improved.